+++ lib/navigation-page.php
@@ -58,12 +58,10 @@ function gutenberg_navigation_init( $hook ) {
 	}
 
 	$settings = array(
-		'disableCustomColors'    => get_theme_support( 'disable-custom-colors' ),
-		'disableCustomFontSizes' => get_theme_support( 'disable-custom-font-sizes' ),
-		'imageSizes'             => $available_image_sizes,
-		'isRTL'                  => is_rtl(),
-		'maxUploadFileSize'      => $max_upload_size,
-		'blockNavMenus'          => get_theme_support( 'block-nav-menus' ),
+		'imageSizes'        => $available_image_sizes,
+		'isRTL'             => is_rtl(),
+		'maxUploadFileSize' => $max_upload_size,
+		'blockNavMenus'     => get_theme_support( 'block-nav-menus' ),
 	);
 
 	list( $color_palette, ) = (array) get_theme_support( 'editor-color-palette' );
@@ -76,6 +74,8 @@ function gutenberg_navigation_init( $hook ) {
 	if ( false !== $font_sizes ) {
 		$settings['fontSizes'] = $font_sizes;
 	}
+
+	$settings = gutenberg_experimental_global_styles_settings( $settings );
 
 	wp_add_inline_script(
 		'wp-edit-navigation',