+++ packages/block-directory/src/plugins/index.js
@@ -2,6 +2,7 @@
  * WordPress dependencies
  */
 import { registerPlugin } from '@wordpress/plugins';
+import { addFilter } from '@wordpress/hooks';
 
 /**
  * Internal dependencies
@@ -9,6 +10,7 @@ import { registerPlugin } from '@wordpress/plugins';
 import AutoBlockUninstaller from '../components/auto-block-uninstaller';
 import InserterMenuDownloadableBlocksPanel from './inserter-menu-downloadable-blocks-panel';
 import InstalledBlocksPrePublishPanel from './installed-blocks-pre-publish-panel';
+import filterMissing from './filter-missing';
 
 registerPlugin( 'block-directory', {
 	render() {
@@ -21,3 +23,9 @@ registerPlugin( 'block-directory', {
 		);
 	},
 } );
+
+addFilter(
+	'editor.missingEdit',
+	'block-directory/install-missing',
+	filterMissing
+);