+++ packages/block-editor/src/components/inserter/quick-inserter.js
@@ -226,6 +226,7 @@ function QuickInserter( {
 					onChange={ ( value ) => {
 						setFilterValue( value );
 					} }
+					placeholder={ __( 'Search for a block' ) }
 				/>
 			) }