+++ packages/react-native-editor/CHANGELOG.md
@@ -12,6 +12,14 @@ For each user feature we should also add a importance categorization label  to i
 ## Unreleased
 * [**] Reflect changes of slider in block settings immediately.
 
+## 1.37.0
+
+* [**] Add support for rounded style in Image block
+
+## 1.36.0
+
+* [**] [Android] Removed pullquote dev only restriction in Android
+
 ## 1.35.0
 
 * [***] Fixed empty text fields on RTL layout. Now they are selectable and placeholders are visible.